# ModeLink/backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ModeLink API",
    "version": "0.1.0",
    "description": "Backend API для проекта ModeLink — цифровая экосистема управления моделями.\n\nАвторизация: Bearer JWT (header Authorization: Bearer <token>)."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {}
        },
        "required": [
          "error",
          "message"
        ]
      },
...
      "ModelProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "height": {
            "type": "integer",
            "nullable": true
          },
          "weight": {
            "type": "integer",
            "nullable": true
          },
          "bust": {
            "type": "integer",
            "nullable": true
          },
          "waist": {
            "type": "integer",
            "nullable": true
          },
          "hips": {
            "type": "integer",
            "nullable": true
          },
          "shoeSize": {
            "type": "number",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "fullName",
          "createdAt",
          "updatedAt"
        ]
      },
      "ModelProfileUpdate": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "height": {
            "type": "integer",
            "minimum": 50,
            "maximum": 250
          },
          "weight": {
            "type": "integer",
            "minimum": 20,
            "maximum": 250
          },
          "bust": {
            "type": "integer",
            "minimum": 40,
            "maximum": 200
          },
          "waist": {
            "type": "integer",
            "minimum": 40,
            "maximum": 200
          },
          "hips": {
            "type": "integer",
            "minimum": 40,
            "maximum": 250
          },
          "shoeSize": {
            "type": "number",
            "minimum": 20,
            "maximum": 60
          }
        }
      },
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "modelId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "modelId",
          "title",
          "date",
          "location",
          "createdAt"
        ]
      },
...
      "PortfolioImage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "modelId": {
            "type": "string",
            "format": "uuid"
          },
          "imageUrl": {
            "type": "string",
            "format": "uri"
          },
          "storageKey": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "modelId",
          "imageUrl",
          "createdAt"
        ]
      },
      "ModelsListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userEmail": {
            "type": "string",
            "format": "email"
          },
          "fullName": {
            "type": "string"
          },
          "height": {
            "type": "integer",
            "nullable": true
          },
          "weight": {
            "type": "integer",
            "nullable": true
          },
          "bust": {
            "type": "integer",
            "nullable": true
          },
          "waist": {
            "type": "integer",
            "nullable": true
          },
          "hips": {
            "type": "integer",
            "nullable": true
          },
          "shoeSize": {
            "type": "number",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "portfolioCount": {
            "type": "integer"
          },
          "projectsCount": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "userEmail",
          "fullName",
          "createdAt",
          "updatedAt",
          "portfolioCount",
          "projectsCount"
        ]
      },
      "ModelDetails": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userEmail": {
            "type": "string",
            "format": "email"
          },
          "fullName": {
            "type": "string"
          },
          "height": {
            "type": "integer",
            "nullable": true
          },
          "weight": {
            "type": "integer",
            "nullable": true
          },
          "bust": {
            "type": "integer",
            "nullable": true
          },
          "waist": {
            "type": "integer",
            "nullable": true
          },
          "hips": {
            "type": "integer",
            "nullable": true
          },
          "shoeSize": {
            "type": "number",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "portfolioImages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortfolioImage"
            }
          },
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Project"
            }
          }
        },
        "required": [
          "id",
          "userEmail",
          "fullName",
          "createdAt",
          "updatedAt",
          "portfolioImages",
          "projects"
        ]
      },
      "PagedModelsList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelsListItem"
            }
          },
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 20
          },
          "total": {
            "type": "integer",
            "example": 100
          },
          "totalPages": {
            "type": "integer",
            "example": 5
          }
        },
        "required": [
          "items",
          "page",
          "limit",
          "total",
          "totalPages"
        ]
      },
...
    }
  },
  "paths": {
...
    "/model-profile/me": {
      "get": {
        "tags": [
          "ModelProfile"
        ],
        "summary": "Профиль модели (текущего пользователя)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelProfile"
                }
              }
            }
          },
          "401": {
            "description": "Нет/неверный JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Роль не model",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/model-profile": {
      "put": {
        "tags": [
          "ModelProfile"
        ],
        "summary": "Обновить профиль модели",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelProfileUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelProfile"
                }
              }
            }
          },
          "401": {
            "description": "Нет/неверный JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Роль не model",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/models": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "List models (search/pagination/sorting)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelsList"
                }
              }
            }
          },
          "401": {
            "description": "Нет/неверный JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Роль не booker/manager/admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "required": false
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            },
            "required": false
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "updatedAt",
                "fullName",
                "portfolioCount",
                "projectsCount"
              ]
            },
            "required": false
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "required": false
          }
        ]
      }
    },
    "/models/{id}": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "Карточка модели (профиль + портфолио + проекты)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDetails"
                }
              }
            }
          },
          "401": {
            "description": "Нет/неверный JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Роль не booker/manager/admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Модель не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/portfolio": {
      "get": {
        "tags": [
          "Portfolio"
        ],
        "summary": "Портфолио: model — свои фото; booker/manager/admin — по modelId",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "modelId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Требуется для booker/manager/admin. Для role=model игнорируется."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PortfolioImage"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Нужен modelId (для booker/manager/admin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Нет/неверный JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Portfolio"
        ],
        "summary": "Загрузить фото в портфолио (multipart/form-data)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Idempotency key for safe retries of portfolio upload (UUID recommended)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "modelId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Требуется для manager/admin. Для role=model игнорируется."
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "image"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioImage"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации или лимит 10 фото",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Нет/неверный JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Роль не model/manager/admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency конфликт (reuse/in-progress)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/portfolio/{id}": {
      "delete": {
        "tags": [
          "Portfolio"
        ],
        "summary": "Удалить фото из портфолио",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Нет/неверный JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Запрещено (модель пытается удалить чужое фото)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Фото не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}